Rename widgets directory to components
closes #817
