+++ app/src/main/java/com/enaboapps/switchify/components/Picker.kt
@@ -1,4 +1,4 @@
-package com.enaboapps.switchify.widgets
+package com.enaboapps.switchify.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
