+++ app/src/main/java/com/enaboapps/switchify/components/PreferenceSwitch.kt
@@ -1,4 +1,4 @@
-package com.enaboapps.switchify.widgets
+package com.enaboapps.switchify.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
