+++ app/src/main/java/com/enaboapps/switchify/screens/EnableAccessibilityServiceScreen.kt
@@ -17,9 +17,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.enaboapps.switchify.R
+import com.enaboapps.switchify.components.FullWidthButton
+import com.enaboapps.switchify.components.NavBar
 import com.enaboapps.switchify.service.utils.ServiceUtils
-import com.enaboapps.switchify.widgets.FullWidthButton
-import com.enaboapps.switchify.widgets.NavBar
 
 @Composable
 fun EnableAccessibilityServiceScreen(navController: NavController) {
