+++ app/src/main/java/com/enaboapps/switchify/screens/account/SignUpScreen.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.enaboapps.switchify.auth.AuthManager
 import com.enaboapps.switchify.auth.GoogleAuthHandler
+import com.enaboapps.switchify.components.*
 import com.enaboapps.switchify.preferences.PreferenceManager
 import com.enaboapps.switchify.service.custom.actions.store.ActionStore
-import com.enaboapps.switchify.widgets.*
 import kotlinx.coroutines.launch
 
 @Composable
