+++ app/src/main/java/com/enaboapps/switchify/screens/howto/HowToUseScreen.kt
@@ -13,11 +13,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import com.enaboapps.switchify.components.InfoCard
+import com.enaboapps.switchify.components.NavBar
+import com.enaboapps.switchify.components.NavRouteLink
+import com.enaboapps.switchify.components.Section
 import com.enaboapps.switchify.nav.NavigationRoute
-import com.enaboapps.switchify.widgets.InfoCard
-import com.enaboapps.switchify.widgets.NavBar
-import com.enaboapps.switchify.widgets.NavRouteLink
-import com.enaboapps.switchify.widgets.Section
 
 @Composable
 fun HowToUseScreen(navController: NavController) {
