+++ app/src/main/java/com/enaboapps/switchify/screens/settings/actions/inputs/SendEmailExtraInput.kt
@@ -2,8 +2,8 @@ package com.enaboapps.switchify.screens.settings.actions.inputs
 
 import android.util.Patterns
 import androidx.compose.runtime.Composable
+import com.enaboapps.switchify.components.TextArea
 import com.enaboapps.switchify.service.custom.actions.store.data.ActionExtra
-import com.enaboapps.switchify.widgets.TextArea
 
 @Composable
 fun SendEmailExtraInput(
