+++ app/src/main/java/com/enaboapps/switchify/screens/settings/menu/MenuItemCustomizationScreen.kt
@@ -9,11 +9,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import com.enaboapps.switchify.components.NavBar
+import com.enaboapps.switchify.components.PreferenceSwitch
 import com.enaboapps.switchify.preferences.PreferenceManager
 import com.enaboapps.switchify.service.menu.MenuItem
 import com.enaboapps.switchify.service.menu.structure.MenuStructureHolder
-import com.enaboapps.switchify.widgets.NavBar
-import com.enaboapps.switchify.widgets.PreferenceSwitch
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
