+++ app/src/main/java/com/enaboapps/switchify/screens/settings/scanning/ScanColorSelectionScreen.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.MutableLiveData
 import androidx.navigation.NavController
+import com.enaboapps.switchify.components.NavBar
 import com.enaboapps.switchify.service.scanning.ScanColorManager
-import com.enaboapps.switchify.widgets.NavBar
 
 @Composable
 fun ScanColorSelectionScreen(
