+++ app/src/main/java/com/enaboapps/switchify/service/custom/actions/AppLaunchPicker.kt
@@ -3,8 +3,8 @@ package com.enaboapps.switchify.service.custom.actions
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
+import com.enaboapps.switchify.components.Picker
 import com.enaboapps.switchify.utils.AppLauncher
-import com.enaboapps.switchify.widgets.Picker
 
 @Composable
 fun AppLaunchPicker(
